+++ pkgs/code_builder/lib/src/specs/control.dart
@@ -9,6 +9,7 @@ import 'package:meta/meta.dart';
 import '../base.dart';
 import 'code.dart';
 import 'expression.dart';
+import 'reference.dart';
 
 part 'control.g.dart';
 
@@ -177,7 +178,7 @@ abstract class WhileLoop
 @immutable
 abstract mixin class ControlTree implements Code, Spec {
   /// The items in this tree.
-  BuiltList<ControlBlock> get blocks;
+  List<ControlBlock?> get _blocks;
 
   @override
   R accept<R>(covariant ControlBlockVisitor<R> visitor, [R? context]) =>
@@ -222,6 +223,9 @@ abstract class Condition
   IfTree get asTree => IfTree.of([this]);
 }
 
+/// Builds a [Condition].
+///
+/// {@category controlFlow}
 abstract class ConditionBuilder
     implements Builder<Condition, ConditionBuilder> {
   ConditionBuilder._();
@@ -302,8 +306,10 @@ abstract class IfTree with ControlTree implements Built<IfTree, IfTreeBuilder> {
       ..insert(0, first);
   }
 
-  @override
   BuiltList<Condition> get blocks;
+
+  @override
+  List<ControlBlock> get _blocks => blocks.toList();
 
   /// Returns a new [IfTree] with [condition] added.
   IfTree withCondition(Condition condition) =>
@@ -324,6 +330,8 @@ abstract class IfTree with ControlTree implements Built<IfTree, IfTreeBuilder> {
 }
 
 /// Builds an [IfTree].
+///
+/// {@category controlFlow}
 abstract class IfTreeBuilder implements Builder<IfTree, IfTreeBuilder> {
   IfTreeBuilder._();
   factory IfTreeBuilder() = _$IfTreeBuilder;
@@ -363,6 +371,155 @@ abstract class IfTreeBuilder implements Builder<IfTree, IfTreeBuilder> {
           body.addExpression(expression.thrown);
         },
       );
+}
+
+/// Represents a `catch` block.
+///
+/// {@category controlFlow}
+abstract class CatchBlock
+    with ControlBlock
+    implements Built<CatchBlock, CatchBlockBuilder> {
+  CatchBlock._();
+  factory CatchBlock([void Function(CatchBlockBuilder) updates]) = _$CatchBlock;
+
+  /// The optional type of exception to catch (`on` clause).
+  ///
+  /// ``` dart
+  /// on type catch (exception)
+  /// on type catch (exception, stacktrace)
+  /// ```
+  Reference? get type;
+
+  /// The name of the exception parameter (default: `e`).
+  ///
+  /// ```dart
+  /// catch (exception)
+  /// catch (exception, stacktrace)
+  /// ```
+  String get exception;
+
+  /// The optional name of the stacktrace parameter.
+  ///
+  /// Will be excluded if left `null`.
+  ///
+  /// ```dart
+  /// catch (exception)
+  /// catch (exception, stacktrace)
+  /// ```
+  String? get stacktrace;
+
+  ControlExpression get _catch =>
+      ControlExpression.catchStatement(exception, stacktrace);
+
+  @override
+  ControlExpression get _expression =>
+      type == null ? _catch : ControlExpression.onStatement(type!, _catch);
+
+  /// Set the default value of [exception]
+  @BuiltValueHook(initializeBuilder: true)
+  static void _initialize(CatchBlockBuilder builder) => builder.exception = 'e';
+}
+
+/// Represents a `try` or `finally` block.
+///
+/// **INTERNAL ONLY**.
+@internal
+class TryBlock with ControlBlock {
+  @override
+  final Block body;
+  final bool isFinally;
+
+  const TryBlock._(this.body) : isFinally = false;
+  const TryBlock._finally(this.body) : isFinally = true;
+
+  @override
+  ControlExpression get _expression => isFinally
+      ? ControlExpression.finallyStatement
+      : ControlExpression.tryStatement;
+}
+
+/// Represents a `try`/`catch` block.
+///
+/// {@category controlFlow}
+abstract class TryCatch
+    with ControlTree
+    implements Built<TryCatch, TryCatchBuilder> {
+  TryCatch._();
+
+  /// Build a [TryCatch].
+  factory TryCatch([void Function(TryCatchBuilder) updates]) = _$TryCatch;
+
+  /// The body of the `try` clause.
+  ///
+  /// ```dart
+  /// try {
+  ///   body
+  /// }
+  /// ```
+  Block get body;
+
+  /// The `catch` clauses for this block.
+  BuiltList<CatchBlock> get handlers;
+
+  /// The optional `finally` clause body.
+  ///
+  /// ```dart
+  /// finally {
+  ///   handleAll
+  /// }
+  /// ```
+  Block? get handleAll;
+
+  TryBlock get _try => TryBlock._(body);
+  TryBlock? get _finally =>
+      handleAll == null ? null : TryBlock._finally(handleAll!);
+
+  @override
+  List<ControlBlock?> get _blocks => [_try, ...handlers, _finally];
+
+  /// Ensure [handlers] is not empty
+  @BuiltValueHook(finalizeBuilder: true)
+  static void _build(TryCatchBuilder builder) =>
+      builder.handlers.isNotEmpty ||
+      (throw ArgumentError(
+          'One or more `catch` clauses must be specified.', 'handlers'));
+}
+
+/// Builds a [TryCatch] block.
+///
+/// {@category controlFlow}
+abstract class TryCatchBuilder implements Builder<TryCatch, TryCatchBuilder> {
+  TryCatchBuilder._();
+  factory TryCatchBuilder() = _$TryCatchBuilder;
+
+  /// The body of the `try` clause.
+  ///
+  /// ```dart
+  /// try {
+  ///   body
+  /// }
+  /// ```
+  BlockBuilder body = BlockBuilder();
+
+  /// The optional `finally` clause body.
+  ///
+  /// ```dart
+  /// finally {
+  ///   handleAll
+  /// }
+  /// ```
+  BlockBuilder? handleAll;
+
+  /// The `catch` clauses for this block.
+  ListBuilder<CatchBlock> handlers = ListBuilder();
+
+  /// Build a `catch` clause and add it to [handlers].
+  void addCatch(void Function(CatchBlockBuilder block) builder) =>
+      handlers.add((CatchBlockBuilder()..update(builder)).build());
+
+  /// Build a `finally` clause and update [handleAll].
+  void addFinally(void Function(BlockBuilder body) builder) =>
+      handleAll = BlockBuilder()..update(builder);
 }
 
 /// Knowledge of different types of control blocks.
@@ -419,7 +576,7 @@ abstract mixin class ControlBlockEmitter
   StringSink visitControlTree(ControlTree tree, [StringSink? output]) {
     output ??= StringBuffer();
 
-    for (final item in tree.blocks) {
+    for (final item in tree._blocks.nonNulls) {
       item.accept(this, output);
       output.write(' ');
     }
@@ -427,7 +584,7 @@ abstract mixin class ControlBlockEmitter
     return output;
   }
 
-@override
+  @override
   StringSink visitControlExpression(ControlExpression expression,
       [StringSink? output]) {
     output ??= StringBuffer();
@@ -483,5 +640,4 @@ abstract mixin class ControlBlockEmitter
 
     return output;
   }
-
 }